refactor: move chain-state and chain-trust globals out of main.cpp (#3125 C9)
Create src/chain.cpp with the definitions the A6 header extraction left
behind in main.cpp: cs_main, the GRC::BlockIndexPool statics,
mapBlockIndex, pindexGenesisBlock, nBestHeight, hashBestChain,
pindexBest, the sync clocks (g_previous_block_time, g_nTimeBestReceived),
g_reorg_in_progress, and CBlockIndex::GetBlockTrust (declared in
primitives/block.h).

Create src/gridcoin/staking/chain_trust.cpp with g_chain_trust,
g_seen_stakes, and GetChainTrust; their canonical declarations move to
gridcoin/staking/chain_trust.h and spam.h (next to the types), replacing
six per-TU ad-hoc externs (node/chainman.cpp, node/coherence.cpp,
node/orphan_blocks.cpp, node/blockstorage.cpp, validation.cpp) and the
GetChainTrust redeclaration in rpc/blockchain.cpp.

Also delete the stale `extern arith_uint256 nBestChainTrust` from
chain.h: it has had no definition anywhere in the tree (the chain-tip
trust lives in g_chain_trust), so any consumer would already fail to
link.

Pure code motion otherwise; no behavior change. Both new .cpp files are
registered in the core source list.
