+++ src/rpc/blockchain.cpp
@@ -18,6 +18,7 @@
 #include <util/string.h>
 #include <util/time.h>
 #include "gridcoin/mrc.h"
+#include "gridcoin/staking/chain_trust.h"
 #include "gridcoin/support/block_finder.h"
 #include "gridcoin/support/xml.h"
 #include <rpc/util.h>
@@ -38,7 +39,6 @@ extern GRC::Superblock ScraperGetSuperblockContract(bool bStoreConvergedStats = 
 extern ScraperPendingBeaconMap GetPendingBeaconsForReport();
 extern ScraperPendingBeaconMap GetVerifiedBeaconsForReport(bool from_global = false);
 extern UniValue GetJSONVersionReport(const int64_t lookback, const bool full_version) EXCLUSIVE_LOCKS_REQUIRED(cs_main);
-arith_uint256 GetChainTrust(const CBlockIndex* pindex) EXCLUSIVE_LOCKS_REQUIRED(cs_main);
 double CoinToDouble(double surrogate);
 extern void TxToJSON(const CTransaction& tx, const uint256 hashBlock, UniValue& entry) EXCLUSIVE_LOCKS_REQUIRED(cs_main);
 UniValue ContractToJson(const GRC::Contract& contract);
